site/requirements.txt
streamlit
google_cloud_firestore